+++ CHANGELOG.md
@@ -16,6 +16,7 @@
 - Ignore deleted and old phone numbers when generating overdue CSV from database
 - Remove extra margin start and end for overdue search button
 - Show selected overdue appointments count in overdue search screen
+- Implement `OverdueAppointmentSelector`
 - [In Progress: 27 Apr 2022] Migrate `BloodPressureHistoryScreenEffectHandler` to use view effect
 
 ## 2022-07-11-8325